FHIR pneumothorax example
Added reference to imagingStudy to the report resource itself and introduced AIM as the newly formed annotation resource along with an observation resource.  This allows for stand alone markups as well as stand alone observations to be made as each can be considered annotations.  Ensured that all RadElemnts in the DSI use case were represented in the example.